+++ src/build_options.rs
@@ -24,14 +24,16 @@ pub struct BuildOptions {
     /// will be rejected by pypi, unless they are separately validated by
     /// `auditwheel`.
     ///
+    /// The default is the lowest supported value for a target, which is 2010 for x86 and 2014 for
+    /// arm and powerpc.
+    ///
     /// This option is ignored on all non-linux platforms
     #[structopt(
         long,
         possible_values = &["2010", "2014", "off"],
         case_insensitive = true,
-        default_value = "2010"
     )]
-    pub manylinux: Manylinux,
+    pub manylinux: Option<Manylinux>,
     #[structopt(short, long)]
     /// The python versions to build wheels for, given as the names of the
     /// interpreters. Uses autodiscovery if not explicitly set.
@@ -77,7 +79,7 @@ pub struct BuildOptions {
 impl Default for BuildOptions {
     fn default() -> Self {
         BuildOptions {
-            manylinux: Manylinux::Manylinux2010,
+            manylinux: None,
             interpreter: Some(vec![]),
             bindings: None,
             manifest_path: PathBuf::from("Cargo.toml"),
@@ -161,6 +163,9 @@ impl BuildOptions {
         }
 
         let target = Target::from_target_triple(self.target.clone())?;
+        let manylinux = self
+            .manylinux
+            .unwrap_or_else(|| target.get_default_manylinux_tag());
 
         let wheel_dir = match self.out {
             Some(ref dir) => dir.clone(),
@@ -191,7 +196,7 @@ impl BuildOptions {
             release,
             strip,
             skip_auditwheel: self.skip_auditwheel,
-            manylinux: self.manylinux,
+            manylinux,
             cargo_extra_args,
             rustc_extra_args,
             interpreter,