+++ Changelog.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (for the cli, not for the crate).
 
+## Unreleased
+
+* Change manylinux default version based on target arch by messense in [#424](https://github.com/PyO3/maturin/pull/424)
+
 ## 0.9.4 - 2021-02-18
 
 * Fix building a bin with musl